fix(GH1687): register live_simulation marker and companion refactor artefacts
Add missing `live_simulation` (and 4 other undeclared) markers to pytest.ini
so that `--strict-markers` no longer raises a collection error on
tests/heavy_integration/test_tools_contracts.py.

Also commit companion refactor files that were already in the working tree
and required to keep ruff clean:
- text_editor_history_mixin.py / text_editor_validation_mixin.py (editor split)
- gas_flow_mixin.py / heating_value_mixin.py / tar_concentration_mixin.py
  (UnitConversionService mixin extraction)

Closes #1687
