+++ pytest.ini
@@ -72,6 +72,12 @@ markers =
     contract: mark test as contract test (validates API/builder contracts without DWSIM runtime)
     acceptance: mark test as acceptance test (end-to-end scenario validation)
     dwsim: mark test as requiring DWSIM runtime (skipped without DWSIM installation)
+    live_simulation: mark test as heavy live-simulation test (weekly runner only, excluded from standard CI)
+    benchmark: mark test as benchmark (measures timing/throughput)
+    scientific: mark test as scientific validation test (numerical accuracy against reference data)
+    headless_safe: mark test as safe to run in headless (no display) environments
+    requires_gl: mark test as requiring OpenGL/GPU display context
+    parity: mark test as cross-repo parity check (validates Tools/UpstreamDrift/Gasification_Model alignment)
 
 # =============================================================================
 # Output and Reporting
